Modifier/afficher les données médicales des patients (dpi/maladie)

+++ back/src/main/java/esiag/back/models/medical/FileAttente.java
@@ -12,7 +12,7 @@ public class FileAttente {
 
     @Id
     @Column(name="id_patient")
-    private int idPatient;
+    private Long idPatient;
 
     @Column(name = "rang")
     private int rang;
